% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_return_dbl_coded.R
\name{int_return_dbl_coded}
\alias{int_return_dbl_coded}
\title{int_return_dbl_coded}
\usage{
int_return_dbl_coded(
  in_object_name,
  in_rater_column,
  in_subject_column,
  in_coding_column
)
}
\arguments{
\item{in_object_name}{A dataframe or tibble containing raters' codings. Each row should contain the assigned coding
from a given rater-subject.}

\item{in_rater_column}{The name of the column containing the raters' names as a string.}

\item{in_subject_column}{The name of the column containing the names of the subjects being coded as a string.}

\item{in_coding_column}{The name of the column containing the codings assigned by the raters as a string.}
}
\description{
\code{int_return_dbl_coded} An internal function to return the subjects double-coded by the raters. It runs a number of checks along the way
}
\author{
Benjamin Goehring <bengoehr@umich.edu>
}
