% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr_stats.R
\name{irr_stats}
\alias{irr_stats}
\title{irr_stats}
\usage{
irr_stats(
  object_name,
  rater_column,
  subject_column,
  coding_column,
  round_digits = 2,
  stats_to_include = c("Percentage agreement", "Cohen's Kappa", "Maxwell's RE",
    "Krippendorf's Alpha")
)
}
\arguments{
\item{object_name}{A dataframe or tibble containing raters' codings. Each row should contain the assigned coding
from a given rater-subject.}

\item{rater_column}{The name of the column containing the raters' names as a string.}

\item{subject_column}{The name of the column containing the names of the subjects being coded as a string.}

\item{coding_column}{The name of the column containing the codings assigned by the raters as a string.}

\item{round_digits}{The number of decimals to round the irr values by. The default is 2.}

\item{stats_to_include}{The IRR statistics to include in the output. See the documentation of the \href{https://cran.r-project.org/web/packages/irr/irr.pdf}{irr package} for more information about the specific statistics.}
}
\description{
\code{irr_stats} Calculates a variety of IRR statistics.
}
\author{
Benjamin Goehring <bengoehr@umich.edu>
}
