% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_for_irr.R
\name{read_for_irr}
\alias{read_for_irr}
\title{read_for_irr}
\usage{
read_for_irr(
  object_name,
  rater_column,
  subject_column,
  coding_column,
  round_digits = 2
)
}
\arguments{
\item{object_name}{A dataframe or tibble containing raters' codings. Each row should contain the assigned coding
from a given rater-subject.}

\item{rater_column}{The name of the column containing the raters' names as a string.}

\item{subject_column}{The name of the column containing the names of the subjects being coded as a string.}

\item{coding_column}{The name of the column containing the codings assigned by the raters as a string.}

\item{round_digits}{The number of decimals to round the irr values by. The default is 2.}
}
\description{
\code{read_for_irr} reads in raters' codings for further analysis from a variety
of formats
}
\author{
Benjamin Goehring <bengoehr@umich.edu>
}
