% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_multicoded.R
\name{return_multicoded}
\alias{return_multicoded}
\title{return_multicoded}
\usage{
return_multicoded(object_name, rater_column, subject_column, coding_column)
}
\arguments{
\item{object_name}{A dataframe or tibble containing raters' codings. Each row should contain the assigned coding
from a given rater-subject.}

\item{rater_column}{The name of the column containing the raters' names as a string.}

\item{subject_column}{The name of the column containing the names of the subjects being coded as a string.}

\item{coding_column}{The name of the column containing the codings assigned by the raters as a string.}
}
\value{
A matrix containing the coding of a given subject for a generic 'coder 1' and 'coder 2'. Coders 1 and 2 do not represent actual coders (e.g., you might have >2). Rather, they are general placeholders used to compare reliability across double-coded subjects.
}
\description{
\code{return_multicoded} Returns a tibble containing the subjects that are
coded by >1 rater and thus able to be used for IRR analysis.
}
\author{
Benjamin Goehring <bengoehr@umich.edu>
}
