% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ura-package.R
\docType{package}
\name{ura-package}
\alias{ura}
\alias{ura-package}
\title{ura: Monitoring Rater Reliability}
\description{
Provides researchers with a simple set of diagnostic tools for monitoring the progress and reliability of raters conducting content coding tasks. Goehring (2024) \url{https://bengoehring.github.io/improving-content-analysis-tools-for-working-with-undergraduate-research-assistants.pdf} argues that supervisors---especially supervisors of small teams---should utilize computational tools to monitor reliability in real time. As such, this package provides easy-to-use functions for calculating inter-rater reliability statistics and measuring the reliability of one coder compared to the rest of the team.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bengoehring/ura}
  \item Report bugs at \url{https://github.com/bengoehring/ura/issues}
}

}
\author{
\strong{Maintainer}: Benjamin Goehring \email{bengoehr@umich.edu} (\href{https://orcid.org/0000-0001-8450-6065}{ORCID})

}
\keyword{internal}
